Adding local sample weights to Ensembles.
We have two options for calculating OOB variable importance a la Breiman
2001:

  1. Calculate in-situ while building the trees.
  2. Save the in-bag bootstrapped sample used while building the trees.

I picked option 2.  This means that the training time is still just as
fast, but the variable importance will not be stored with the trees (and
will therefore need to be re-calculated every time they are used.)  I
think this is okay because typical use would be to either plot the
variable importances once or save them to some local variable for
processing.

Note that this also adds a matrix of size NxM to the Ensemble, where N
is the number of samples used to train the Ensemble and M is the number
of estimators trained.  I fear this will have to be revisited as N might
be extremely large.